+++ Engine/Title.cpp
@@ -18,52 +18,52 @@ void Title::Init(std::mt19937& rng)
 
 void Title::DrawRegularStar(Graphics& gfx)
 {
-	art.Regularstar(loc.x, loc.y, gfx);
+	art.Regularstar((int)loc.x, (int)loc.y, gfx);
 }
 
 void Title::DrawDimStar(Graphics& gfx)
 {
-	art.DimStar(loc.x, loc.y, gfx);
+	art.DimStar((int)loc.x, (int)loc.y, gfx);
 }
 
 void Title::DrawBrightStar(Graphics& gfx)
 {
-	art.BrightStar(loc.x, loc.y, gfx);
+	art.BrightStar((int)loc.x, (int)loc.y, gfx);
 }
 
 void Title::DrawTitle(Graphics& gfx)
 {
-	art.DrawTitle(loc.x, loc.y, gfx);
+	art.DrawTitle((int)loc.x, (int)loc.y, gfx);
 }
 
 void Title::DrawPrview(Graphics& gfx)
 {
-	art.Drawtitlepreview(loc.x - 100, loc.y + 200, gfx);
+	art.Drawtitlepreview((int)loc.x - 100, (int)loc.y + 200, gfx);
 }
 
 void Title::DrawYoda(Graphics& gfx)
 {
-	art.YodaArt(loc.x - 180, loc.y, gfx);
+	art.YodaArt((int)loc.x - 180, (int)loc.y, gfx);
 }
 
 void Title::DrawYodaSpeaks(Graphics& gfx)
 {
-	art.YodaSpeaks(loc.x, loc.y, gfx);
+	art.YodaSpeaks((int)loc.x, (int)loc.y, gfx);
 }
 
 void Title::DrawJediMenu(Graphics& gfx)
 {
-	art.Jedichoice(loc.x - 280, loc.y + 200, gfx);
+	art.Jedichoice((int)loc.x - 280, (int)loc.y + 200, gfx);
 }
 
 void Title::DrawSaberMenu(Graphics& gfx)
 {
-	art.Saberchoice(loc.x + 50, loc.y + 200 , gfx);
+	art.Saberchoice((int)loc.x + 50, (int)loc.y + 200 , gfx);
 }
 
 void Title::DrawSaberSelect0r1(Graphics& gfx)
 {
-	art.SelectingSaber1(loc.x, loc.y, gfx);
+	art.SelectingSaber1((int)loc.x, (int)loc.y, gfx);
 }
